Remove dependencies from storyshots
Storyshots wants to plugin to the version used by the app, not have its own version. So it should probably use a peer dep; however as some deps are optional (i.e. react vs react-native) this will lead to confusing warnings to the user.

This seems like an OK approach until we hear differently from users.